+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ---
 
 ### New
+* Enable toggling aborting CodeBuild phases on command execution failure
+* Support CodeBuild spec runtime-versions
 
 ### Changes